refactor: rename memory to volitale
