+++ catalog_impls/src/lib.rs
@@ -7,9 +7,9 @@ use system_catalog::{tables::Tables, SystemTableAdapter};
 
 use crate::system_tables::{SystemTables, SystemTablesBuilder};
 
-pub mod memory;
 mod system_tables;
 pub mod table_based;
+pub mod volitale;
 
 /// CatalogManagerImpl is a wrapper for system and user tables
 #[derive(Clone)]
